ai-agents/stock-agent/requirements.txt
claude-agent-sdk==0.1.3
python-dotenv==1.0.0
